Rename auth feature to login
