+++ lib/services/api/api_client.dart
@@ -1,6 +1,6 @@
 import 'package:dio/dio.dart';
 
-import '../../features/auth/models/login.dart';
+import '../../features/login/models/login.dart';
 import '../../features/posts/models/post.dart';
 import '../../features/products/models/product.dart';
 import '../../features/profile/models/profile.dart';
